+++ CHANGELOG.md
@@ -11,9 +11,11 @@ and this project adheres to [Semantic Versioning].
 
 - Basic project structure.
 - `Matrix` class to represent images.
-- Implement `DisparityGraph` class to store images
+- `DisparityGraph` class to store images
   and calculate penalties for different labelings.
-- Implement `DisparityNode` to request and identify nodes of the graph.
+- `DisparityNode` to request and identify nodes of the graph.
+- `Labeling` to represent, change and calculate penalty
+  of particular labelings.
 
 [Keep a Changelog]: http://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: http://semver.org/spec/v2.0.0.html